% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bf_int.R
\name{bf_int}
\alias{bf_int}
\title{Bonferroni Confidence Intervals for \eqn{\lambda} Values}
\usage{
bf_int(data, q, alpha = 0.05)
}
\arguments{
\item{data}{a data frame}

\item{q}{the number of variances we want to find confidence intervals for. We find confidence intervals for the first 1:q variances}

\item{alpha}{the significance level, default set to the standard 0.05}
}
\value{
ci, an object containing the q confidence intervals as a matrix
}
\description{
We use the formula given on pages 456 and 457 of our book \emph{Applied Multivariate Statistical Analysis},
 6th Edition, by Johnson and Wichern, to create our formula for the confidence intervals.
}
\examples{
data <- datasets::mtcars
bf_int(data, q = 2)

}
