% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BFconf.ellipse.R
\name{BFconf.ellipse}
\alias{BFconf.ellipse}
\title{Bonferroni Confidence Intervals with \eqn{T[2]} Comparison}
\usage{
BFconf.ellipse(data, alpha = 0.05, col1 = 1, col2 = 2, addT2 = FALSE)
}
\arguments{
\item{data}{a two-vector data frame}

\item{alpha}{the significance level}

\item{col1}{an integer saying which is the first column the user would like to look at}

\item{col2}{an integer saying which is the second column the user would like to look at}

\item{addT2}{a boolean variable. Default is FALSE but, if TRUE, adds \eqn{T[2]} intervals to the graph}
}
\value{
g a ggplot object containing the confidence ellipse with a dashed line showing the Bonferroni intervals and (if addT2 is set to TRUE) a dotted line showing the \eqn{T[2]} intervals
}
\description{
Bonferroni Confidence Intervals with \eqn{T[2]} Comparison
}
\examples{
testData <- datasets::mtcars
BFconf.ellipse(testData)

}
