% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfMean_int.R
\name{bfMean_int}
\alias{bfMean_int}
\title{(1 - \eqn{\alpha})*100\% Bonferroni Confidence Intervals for \eqn{\mu}}
\usage{
bfMean_int(data, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame}

\item{alpha}{our \emph{overall} significance level, set to the standard 0.05}
}
\value{
The \eqn{(1 - \alpha)*100}\% Bonferroni confidence intervals
}
\description{
The formula used to create this confidence interval cis 5-29 from page 232 of our book, \emph{Applied Multivariate Statistical Analysis},
 6th Edition, by Johnson and Wichern
}
\examples{
testData <- datasets::mtcars
bfMean_int(data = testData)

}
