% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bookSD.R
\name{bookSD}
\alias{bookSD}
\title{Book Standard Deviation}
\usage{
bookSD(x, n)
}
\arguments{
\item{x}{a vector of quantitative data}

\item{n}{the number of observations in the vector}
}
\value{
standard deviation from the biased formula
}
\description{
Book Standard Deviation
}
\examples{
x = 1:10
bookSDX <- bookSD(x, n=10)
}
