% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{box_cox}
\alias{box_cox}
\title{Box-Cox Transformation}
\usage{
box_cox(x, lambda)
}
\arguments{
\item{x}{a vector of data}

\item{lambda}{the value of lambda for the Box-Cox transformation}
}
\value{
x_lamb the transformed data
}
\description{
Returns the transformed values of a particular column of data using the Box-Cox method for a
particular value of lambda. The transformation is \eqn{\frac{x^{\lambda} - 1}{x}} if \eqn{\lambda ≠ 0} and
\eqn{log_e(x)} if \eqn{\lambda = 0}
}
\examples{
box_cox(x = 1:10, lambda = 0)
}
