% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chi_sqQ.R
\name{chi_sqQ}
\alias{chi_sqQ}
\title{Values for Chi-Square Plot}
\usage{
chi_sqQ(df)
}
\arguments{
\item{df}{a bivariate data.frame}
}
\value{
j the observation number, ordered by the squared distance

d_jsq the squared distance

q_j the chi-square quantiles from the chi-square distribution
}
\description{
This function is for use in my shiny_MVN app to calculate the
values for the chi-square plot for assessing normality
}
\examples{
x1 <- c(108.28, 152.36, 95.04, 65.45, 62.97, 263.99, 265.19, 285.06, 92.01, 165.68)
x2 <- c(17.05, 16.59, 10.91, 14.14, 9.52, 25.33, 18.54, 15.73, 8.10, 11.13)
data <- data.frame(cbind(x1, x2))
chi_sqQ(data)
}
