% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ci_T2.R
\name{ci_T2}
\alias{ci_T2}
\title{Simultaneous \eqn{T^2} Confidence Intervals}
\usage{
ci_T2(data, alpha = 0.05)
}
\arguments{
\item{data}{a data.frame}

\item{alpha}{the significance level for the intervals - default to standard 0.05}
}
\value{
the 100(\eqn{\alpha})\% confidence intervals based on \eqn{T^2}
}
\description{
We use the formula given in equation (5-24) on page 224 of our
book \emph{Applied Multivariate Statistical Analysis}, 6th Edition, by Johnson and Wichern, to create our
formula for the confidence intervals.
}
\examples{
testData <- datasets::mtcars[,1:4]
ci_T2(testData, 0.05)
}
