% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conf.ellipse.R
\name{conf.ellipse}
\alias{conf.ellipse}
\title{Confidence regions for bivariate data}
\usage{
conf.ellipse(data, alpha, mu_0)
}
\arguments{
\item{data}{a data.frame containing two variables}

\item{alpha}{a scaled quantity between 0 and 1; 100(1-alpha) is the confidence level}

\item{mu_0}{the vector of means from the null hypothesis}
}
\value{
a plot showing the ellipse that is the confidence region based on alpha

a named list with several outputs, explained further in the output section
}
\description{
Confidence regions for bivariate data
}
\section{List of output}{


The list returned at the end of the function has several components to its output:

\enumerate{
\item The result of the confidence test (result)
\item The size of the quadratic form (quadratic.form)
\item The scaled quantile (scaled.quantile)
\item The eigenvectors and eigenvalues of S (eigenvalues and eigenvectors)
\item The major and minor half-widths (major.halfwidth and minor.halfwidth)
\item The lengths ratio, i.e. ratio of the major axis to the minor axis (lengths.ratio)
}
}

\examples{
\dontrun{
df <- diamonds[, c(1,5)]
dBar <- colMeans(data)
conf.ellipse(data = df, alpha = 0.05, mu_0 = dBar)}
}
