% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_coefqq.R
\name{cor_coefqq}
\alias{cor_coefqq}
\title{Correlation Coefficient}
\usage{
cor_coefqq(x)
}
\arguments{
\item{x}{the column of a data frame}
}
\value{
rq the correlation coefficient for the given column
}
\description{
Correlation Coefficient
}
\examples{
\dontrun{
cor_coefqq(mtcars[,1])
}
}
