% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop1_quants.R
\name{drop1_quants}
\alias{drop1_quants}
\title{Drop-1 Quantiles}
\usage{
drop1_quants(x, dfDropPoints)
}
\arguments{
\item{x}{a column of data}

\item{dfDropPoints}{the observation to be dropped}
}
\value{
orderedObs a vector containing the entered data, in ascending order, with the specified observation dropped

probabilityLevels a vector containing the probability level for each ordered observation, calculated with the specified observation dropped

stndNormQuant the standard normal quantile, used for creating a QQ-Plot, calculated with the specified observation dropped
}
\description{
This function will input
}
\examples{
x <- 1:10
qq <- q_quantiles(x)
qd <- qq$stndNormQuant[1] # the observation to be dropped
xd <- qq$orderedObs[1]
dfDrop <- data.frame(xj = 1325, qj = qd)
drop1_quants(x = x, dfDropPoints = dfDrop)
}
