% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/l_lambda.R
\name{l_lambda}
\alias{l_lambda}
\title{l(\eqn{\lambda}) Values}
\usage{
l_lambda(x, lambda)
}
\arguments{
\item{x}{a column of data}

\item{lambda}{a vector of values for the Box-Cox power transformation}
}
\value{
l_lambda, a vector of the same length as lambda for the corresponding l(\eqn{\lambda}) values
}
\description{
l(\eqn{\lambda}) Values
}
\examples{
x = 1:10
lambda_vector <- seq(0.001, 1, by = 0.001)
l_lamb <- l_lambda(x, lambda_vector)
}
