% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfirstfun.R
\name{myfirstfun}
\alias{myfirstfun}
\title{Title}
\usage{
myfirstfun(x)
}
\arguments{
\item{x}{a vector of quantitative data}
}
\value{
a vector of squared components
}
\description{
Title
}
\examples{
y <- 1:10
myfirstfun(y)
}
