% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_sigma.R
\name{pca_sigma}
\alias{pca_sigma}
\title{PCA of the Sample Covariance}
\usage{
pca_sigma(data)
}
\arguments{
\item{data}{a data frame}
}
\value{
sample.Cov, the sample covariance

PC.Sigma, the principle components (the coefficients)

rho_yixk, the matrix of the correlation between each principle component and each column of data

TSV, the total sample variance
}
\description{
PCA of the Sample Covariance
}
\examples{
data <- datasets::mtcars
pca_sigma(data[,1:3])
}
