% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcabootAG.R
\name{pcabootAG}
\alias{pcabootAG}
\title{PCA and Bootstrapping}
\usage{
pcabootAG(data, alpha, iter, cov = TRUE)
}
\arguments{
\item{data}{a data frame}

\item{alpha}{the significance level}

\item{iter}{the number of iterations to use when bootstrapping}

\item{cov}{a boolean for if we want to use S (cov = TRUE) or R (cov = FALSE) for the analysis}
}
\value{
a barplot for the bootstrapped values of each \eqn{\lambda[i]}

a violin plot for the bootstrapped values for each e_ik, i,k = 1,2,...p, where p is the number of variables

a list containing confidence intervals for each \eqn{\lambda[i]} and e_ik, i,k = 1,2,...p from bootstrapping, and confidence intervals for each \eqn{\lambda[i]} using the Anderson and Gerschick method
}
\description{
This is a function for the PCA bootstrapping assignment. It performs bootstrapping on the
eigenvalues and eigenvectors of either the sample covariance (S) or the sample correlation (R), then it
outputs plots of the bootstrapped estimates, as well as confidence intervals. It also outputs Anderson and
Girschick confidence intervals for the \eqn{\lambda} values
}
\examples{
data <- as.data.frame(datasets::mtcars)
pcabootAG(data[,1:3], alpha = 0.05, iter = 100)


}
