% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q_quantiles.R
\name{q_quantiles}
\alias{q_quantiles}
\title{Data Quantiles}
\usage{
q_quantiles(x)
}
\arguments{
\item{x}{a column of data}
}
\value{
orderedObs a vector containing the entered data, in ascending order

probabilityLevels a vector containing the probability level for each ordered observation

stndNormQuant the standard normal quantile, used for creating a QQ-Plot
}
\description{
Data Quantiles
}
\examples{
x <- c(-1.00, -.10, .16, .41, .62, .80, 1.26, 1.54, 1.71, 2.30)
q_quantiles(x)

}
