% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stnd_values.R
\name{stnd_values}
\alias{stnd_values}
\title{Standardized Score (Z-scores) for a data frame}
\usage{
stnd_values(data)
}
\arguments{
\item{data}{a data.frame of continuous variables}
}
\value{
Z a matrix of the standardized variables for each entry in the data frame
}
\description{
Standardized Score (Z-scores) for a data frame
}
\examples{
stnd_values(mtcars)
}
