% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T2conf.ellipse.R
\name{T2conf.ellipse}
\alias{T2conf.ellipse}
\title{(1 - \eqn{\alpha})*100\% Confidence Ellipses}
\usage{
T2conf.ellipse(data, alpha = 0.05, col1 = 1, col2 = 2)
}
\arguments{
\item{data}{a two-vector data.frame}

\item{alpha}{the significance level, set to the default of 0.05}

\item{col1}{an integer saying which is the first column the user would like to look at}

\item{col2}{an integer saying which is the second column the user would like to look at}
}
\value{
g a ggplot2 object with a drawing of the confidence ellipse (solid) with the T^2 intervals drawn on with dashed lines
}
\description{
I used Figures 5.1 and 5.2 o pages 23 and 227, respectively, of our
book, \emph{Applied Multivariate Statistical Analysis}, 6th Edition, by Johnson and Wichern, to create the confidence
ellipse. This is similar to the funciton I made called conf.ellipse, but this one \emph{only} outputs the graph, it
outputs it as an object (so I can use it in my Shiny app), and it adds the dashed lines
}
\examples{
testData <- datasets::mtcars[, 1:2]
T2conf.ellipse(testData, 0.05)
}
