% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsq_chart.R
\name{tsq_chart}
\alias{tsq_chart}
\title{\eqn{T^2} Chart}
\usage{
tsq_chart(df)
}
\arguments{
\item{df}{a numeric multivariate data.frame}
}
\value{
p a \eqn{T^2} chart for the data frame with a dashed line at the 95% limit and a solid line at the 99% limit
}
\description{
\eqn{T^2} Chart
}
\examples{
tsq_chart(datasets::mtcars)
}
