% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weld.R
\docType{data}
\name{weld}
\alias{weld}
\title{Data on welding tube yokes to a tube for assembling cars}
\format{
A data frame with 40 rows and 4 variables:
\describe{
  \item{Voltage}{Voltage, volts}
  \item{Current}{Current, amps}
  \item{FeedSpeed}{Feed speed, in/min}
  \item{GasFlow}{(intert) Gas flow, cfm}
}
}
\source{
Applied Multivariate Statistical Analysis, 6th Edition, by Johnson and Wichern
}
\usage{
weld
}
\description{
A dataset containing data on 40 observations of four variables from a machine welding tube yokes to a tube
}
\keyword{datasets}
