% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XbarCharts.R
\name{XbarCharts}
\alias{XbarCharts}
\title{Xbar Charts}
\usage{
XbarCharts(x)
}
\arguments{
\item{x}{a column of data}
}
\value{
a graph showing the univariate X-bar chart for x
}
\description{
Xbar Charts
}
\examples{
weldln <- data.frame(cbind(weld[,1:3], log(weld[,4])))
colnames(weldln) <- c("Voltage", "Current", "FeedSpeed", "ln(GasFlow)")
y <- weldln$`ln(GasFlow)`
XbarCharts(y)
}
